Export `removeTrailingSpaces()` as part of utils
To be importable by the apps, as it's needed for implementing fallback verification following the fix in #151
