+++ lib/message/encrypt.js
@@ -1,7 +1,7 @@
 import { isStream, passiveClone, clone } from '@openpgp/web-stream-tools';
 import { generateSessionKey, encrypt, sign, createMessage, armor as openpgp_armor, enums } from '../openpgp';
 import { serverTime } from '../serverTime';
-import { removeTrailingSpaces } from './utils';
+import { removeTrailingSpaces } from '../utils';
 
 export default async function encryptMessage({
     textData,
