+++ lib/message/processMIME.js
@@ -1,4 +1,4 @@
-import { removeTrailingSpaces } from './utils';
+import { removeTrailingSpaces } from '../utils';
 import { verifyMessage } from './verify';
 import { VERIFICATION_STATUS, MAX_ENC_HEADER_LENGTH } from '../constants';
 import { serverTime } from '../serverTime';
