+++ lib/message/sign.js
@@ -1,6 +1,6 @@
 import { createMessage, sign } from '../openpgp';
 import { serverTime } from '../serverTime';
-import { removeTrailingSpaces } from './utils';
+import { removeTrailingSpaces } from '../utils';
 
 /**
  * Get a signed message from the given data.
