+++ lib/message/verify.js
@@ -1,7 +1,7 @@
 import { createMessage, verify, CleartextMessage } from '../openpgp';
 import { VERIFICATION_STATUS } from '../constants';
 import { serverTime } from '../serverTime';
-import { removeTrailingSpaces } from './utils';
+import { removeTrailingSpaces } from '../utils';
 
 const { NOT_SIGNED, SIGNED_AND_VALID, SIGNED_AND_INVALID } = VERIFICATION_STATUS;
