+++ lib/utils.ts
@@ -175,3 +175,17 @@ export function utf8ArrayToString(utf8: MaybeStream<Uint8Array>): MaybeStream<st
 
     return transformedStream;
 }
+
+/**
+ * Remove trailing spaces, carriage returns and tabs from each line (separated by \n characters)
+ */
+export const removeTrailingSpaces = (text: string) => {
+    return text
+        .split('\n')
+        .map((line) => {
+            let i = line.length - 1;
+            for (; i >= 0 && (line[i] === ' ' || line[i] === '\t' || line[i] === '\r'); i--);
+            return line.substring(0, i + 1);
+        })
+        .join('\n');
+};
